relocate file path
